requirements.txt
pygame
PyOpenGL_whl/PyOpenGL_accelerate-3.1.7-cp312-cp312-win_amd64.whl
PyOpenGL_whl/PyOpenGL-3.1.7-cp312-cp312-win_amd64.whl